+++ src/consumer/component/import-components.js
@@ -1,6 +1,7 @@
 /** @flow */
 import R from 'ramda';
 import chalk from 'chalk';
+import fs from 'fs-extra';
 import { NothingToImport } from '../exceptions';
 import { BitId, BitIds } from '../../bit-id';
 import Component from '../component';
@@ -9,7 +10,7 @@ import { ComponentWithDependencies } from '../../scope';
 import loader from '../../cli/loader';
 import { BEFORE_IMPORT_ACTION } from '../../cli/loader/loader-messages';
 import logger from '../../logger/logger';
-import { filterAsync, pathNormalizeToLinux } from '../../utils';
+import { filterAsync, pathNormalizeToLinux, isDir, isDirEmptySync } from '../../utils';
 import GeneralError from '../../error/general-error';
 import type { MergeStrategy, FilesStatus } from '../versions-ops/merge-version/merge-version';
 import { applyModifiedVersion } from '../versions-ops/checkout-version';
@@ -73,6 +74,7 @@ export default class ImportComponents {
     logger.debug(`importSpecificComponents, Ids: ${this.options.ids.join(', ')}`);
     // $FlowFixMe - we check if there are bitIds before we call this function
     const bitIds = this.options.ids.map(raw => BitId.parse(raw));
+    if (this.options.writeToPath) this._throwErrorWhenDirectoryNotEmpty(this.options.writeToPath);
     const beforeImportVersions = await this._getCurrentVersions(bitIds);
     await this._warnForModifiedOrNewComponents(bitIds);
     const componentsWithDependencies = await this.consumer.scope.getManyWithAllVersions(bitIds, false);
@@ -195,6 +197,19 @@ export default class ImportComponents {
       );
     }
     return Promise.resolve();
+  }
+
+  _throwErrorWhenDirectoryNotEmpty(writeToPath: string) {
+    if (fs.pathExistsSync(writeToPath)) {
+      if (!isDir(writeToPath)) {
+        throw new GeneralError(`unable to import to ${writeToPath} because it's a file`);
+      }
+      if (!isDirEmptySync(writeToPath) && !this.options.override) {
+        throw new GeneralError(
+          `unable to import to ${writeToPath}, the directory is not empty. use --override flag to delete the directory and then import`
+        );
+      }
+    }
   }
 
   async _getMergeStatus(componentWithDependencies: ComponentWithDependencies): Promise<ComponentMergeStatus> {